refactoring and plots
